Change: ufDataの変換を別ファイルに移動
