+++ src/store/singing.ts
@@ -78,7 +78,6 @@ import {
   removeNotesFromOverlappingNoteInfos,
   updateNotesOfOverlappingNoteInfos,
 } from "@/sing/storeHelper";
-import { getDoremiFromNoteNumber } from "@/sing/viewHelper";
 import {
   AnimationTimer,
   createPromiseThatResolvesWhen,
@@ -89,6 +88,7 @@ import { getWorkaroundKeyRangeAdjustment } from "@/sing/workaroundKeyRangeAdjust
 import { createLogger } from "@/domain/frontend/log";
 import { noteSchema } from "@/domain/project/schema";
 import { getOrThrow } from "@/helpers/mapHelper";
+import { ufDataToSongState } from "@/sing/ufDataToSongState";
 
 const logger = createLogger("store/singing");
 
@@ -1694,121 +1694,11 @@ export const singingStore = createPartialStore<SingingStoreTypes>({
   IMPORT_EXTERNAL_PROJECT_FILE: {
     action: createUILockAction(
       async ({ state, dispatch }, { project, trackIndex = 0 }) => {
-        const convertPosition = (
-          position: number,
-          sourceTpqn: number,
-          targetTpqn: number,
-        ) => {
-          return Math.round(position * (targetTpqn / sourceTpqn));
-        };
-
-        const convertDuration = (
-          startPosition: number,
-          endPosition: number,
-          sourceTpqn: number,
-          targetTpqn: number,
-        ) => {
-          const convertedEndPosition = convertPosition(
-            endPosition,
-            sourceTpqn,
-            targetTpqn,
-          );
-          const convertedStartPosition = convertPosition(
-            startPosition,
-            sourceTpqn,
-            targetTpqn,
-          );
-          return Math.max(1, convertedEndPosition - convertedStartPosition);
-        };
-
-        const removeDuplicateTempos = (tempos: Tempo[]) => {
-          return tempos.filter((value, index, array) => {
-            return (
-              index === array.length - 1 ||
-              value.position !== array[index + 1].position
-            );
-          });
-        };
-
-        const removeDuplicateTimeSignatures = (
-          timeSignatures: TimeSignature[],
-        ) => {
-          return timeSignatures.filter((value, index, array) => {
-            return (
-              index === array.length - 1 ||
-              value.measureNumber !== array[index + 1].measureNumber
-            );
-          });
-        };
-
-        // 歌詞をひらがなの単独音に変換する
-        // TODO: 手動で変換元を選べるようにする
-        const convertedProject = project.convertJapaneseLyrics(
-          "auto",
-          "KanaCv",
-          { convertVowelConnections: true },
+        const { tempos, timeSignatures, tracks, tpqn } = ufDataToSongState(
+          project.toUfData(),
         );
 
-        // https://github.com/sdercolin/utaformatix-data?tab=readme-ov-file#value-conventions
-        const projectTpqn = 480;
-        const projectTempos = convertedProject.tempos;
-        const projectTimeSignatures = convertedProject.timeSignatures;
-
-        const trackNotes = convertedProject.tracks[trackIndex].notes;
-
-        trackNotes.sort((a, b) => a.tickOn - b.tickOn);
-
-        // utaformatixはフォールバック歌詞を「あ」としている。
-        // このため、歌詞が「あ」の場合は歌詞が設定されていないとみなす。
-        // TODO: false positiveが起きる可能性があるので、他の方法を考える
-        //   最終的にやりたいことは歌詞が無い時に「ド」～「シ」の歌詞を設定することなので、
-        //   本家のフォールバックを変えるPRを送る？
-        // https://github.com/sdercolin/utaformatix3/blob/baee542392421a628d424d8325a5e0f14d0f2a50/src/jsMain/kotlin/model/Constants.kt#L6
-        const hasLyric = trackNotes.some((value) => value.lyric !== "あ");
-
-        const tpqn = DEFAULT_TPQN;
-
-        const notes = trackNotes.map((value): Note => {
-          return {
-            id: NoteId(uuidv4()),
-            position: convertPosition(value.tickOn, projectTpqn, tpqn),
-            duration: convertDuration(
-              value.tickOn,
-              value.tickOff,
-              projectTpqn,
-              tpqn,
-            ),
-            noteNumber: value.key,
-            lyric: hasLyric
-              ? // たまに空文字が入っていることがあるので、その場合は「っ」に変換する
-                value.lyric === ""
-                ? "っ"
-                : value.lyric
-              : getDoremiFromNoteNumber(value.key),
-          };
-        });
-
-        let tempos = projectTempos.map((value): Tempo => {
-          return {
-            position: convertPosition(value.tickPosition, projectTpqn, tpqn),
-            bpm: round(value.bpm, 2),
-          };
-        });
-        tempos.unshift(createDefaultTempo(0));
-        tempos = removeDuplicateTempos(tempos);
-
-        let timeSignatures: TimeSignature[] = [];
-        for (const ts of projectTimeSignatures) {
-          const beats = ts.numerator;
-          const beatType = ts.denominator;
-          timeSignatures.push({
-            measureNumber: ts.measurePosition,
-            beats,
-            beatType,
-          });
-        }
-        timeSignatures.unshift(createDefaultTimeSignature(1));
-        timeSignatures = removeDuplicateTimeSignatures(timeSignatures);
+        const notes = tracks[trackIndex].notes;
 
         tempos.splice(1, tempos.length - 1); // TODO: 複数テンポに対応したら削除
         timeSignatures.splice(1, timeSignatures.length - 1); // TODO: 複数拍子に対応したら削除
